requirements.txt
fastapi==0.110.0
langchain==0.0.184
openai==0.27.6
pypdf==4.1.0
tiktoken==0.6.0
uvicorn==0.28.0
python-dotenv==1.0.0